### 2016-08-28 #55 @whedon
[REVIEW]: RcppCNPy: Read-Write Support for NumPy Files in R
**Submitting author:** @eddelbuettel (<a href="http://orcid.org/0000-0001-6419-907X">Dirk Eddelbuettel</a>)
**Repository:** <a href="https://github.com/eddelbuettel/rcppcnpy" target ="_blank">https://github.com/eddelbuettel/rcppcnpy</a>
**Version:** 0.2.5
