### 2016-08-28 c576ca2
added example to README.md
