### 2016-09-20 #55 @cMadan
@mdnunez I tried to assign you as a reviewer so you can modify the checklist, but that didn't work--will sort this out, but feel free to try the package

@arfon do reviewers need to be part of the joss-reviewers list to be assign-able?
