### 2016-09-20 #55 @arfon
@cMadan - apologies. There was a small bug with who could assign reviewers. I've fixed this now and @mdnunez you should be set up with collaborator access on this repo so you can update the checklist (you don't need to be in the `@openjournals/joss-reviewers`) team any longer.
