### 2016-09-20 #55 @cMadan
@arfon - thanks for sorting that out!

@mdnunez - take a look at the reviewer guidelines here (http://joss.theoj.org/about). Let me know if you have any questions!
