### 2016-09-20 #55 @mdnunez
I will have some comments. Should I leave them here or make a rcppcnpy repo issue?
