### 2016-09-20 #55 @arfon
> I will have some comments. Should I leave them here or make a rcppcnpy repo issue?

@mdnunez if they are related to the review then please leave comments here for @eddelbuettel to respond to.

If they are issues with things such as installation troubles (i.e. things that would be of interest to any user of the package) then please open issues in the https://github.com/eddelbuettel/rcppcnpy repository.
