### 2016-09-20 #55 @eddelbuettel
Hi @mdnunez -- thanks for volunteering to review!  I am subscribed here, so I would see them either way.

From what I gather from the other joss-reviews I have seen here, feedback in this thread seems common.   But as @arfon said, any "structural" issues in the package can go to the repo as well.

In case it helps, the package also has binaries for Windows and OS X via [its CRAN page](https://cloud.r-project.org/web/packages/RcppCNPy/index.html) and the normal CRAN installs.

Looking forward to your comments.
