### 2016-09-20 #55 @mdnunez
# Review of RcppCNPy: Read-Write Support for NumPy Files in R

> Yes, it is also available from CRAN with install.packages('rcppcnpy')

> Yes, except it is GNU 2, not GNU 3 or later. Also, should GNU statement be at the top of each script?

> README does not list installation instructions, easy fix
> library(rcppcnpy) did not work for me
> It seems library('RcppCNPy') command is case sensitive.

> N/A? I did not see any performance claims in README, I found performance claims in https://cran.r-project.org/web/packages/RcppCNPy/vignettes/RcppCNPy-intro.pdf

> Target audience could be better stated in README and vignette

> I'm not sure if the authors would also like to include the list of Python dependencies since Python example code is given in the README

> Functionality is better stated in vignette. README does not contain exhaustive list of functionality

> Tests were performed, but I did not find a .md or README if that is required by this journal

> Community guidelines for contributions could be better reported

> Author names present in paper.md but did not compile to pdf

> References do not have a DOI in pdf

### Recommendations to editor

While the software works well and quickly, some additional documentation is required by JOSS for publication.

### Recommendations to author

This reviewer would like to see reading and writing capability of .npz in the future. npyLoad of integer arrays produces incorrect values when not given the flag "integer." Multidimensional arrays should also be easily loadable. I was impressed that arrays of sizes (24,) , (1,24), and (24,1) all loaded correctly in R with the existing functionality.

### Recommendations to JOSS

Reviewers should be left with the choice to be anonymous
