### 2016-09-21 #55 @eddelbuettel
If the above was accessible in markdown for me, it would make my reply a little easier.

For the time being I am copying-and-pasting-and-reformatting which feel very 1990s to me...
