+++ paper/joss/2016-09-20-review.md
@@ -1,0 +1,109 @@
+# Review of RcppCNPy: Read-Write Support for NumPy Files in R
+
+## Reviewer questions
+
+### Conflict of interest
+
+[x] As the reviewer I confirm that there are no conflicts of interest
+for me to review this work (such as being a major contributor to the software).
+
+### General checks
+
+[x] Repository: Is the source code for this software available at the repository url?
+
+> Yes, it is also available from CRAN with install.packages('rcppcnpy')
+
+[x] License: Does the repository contain a plain-text LICENSE file
+with the contents of an OSI approved software license?
+
+> Yes, except it is GNU 2, not GNU 3 or later. Also, should GNU
+> statement be at the top of each script?
+
+[x] Version: Does the release version given match the GitHub release (0.2.5)?
+
+### Functionality
+
+[ ]  Installation: Does installation proceed as outlined in the documentation?
+
+> README does not list installation instructions, easy fix.  
+`library(rcppcnpy)` did not work for me  
+It seems `library('RcppCNPy')` command is case sensitive.
+
+[x] Functionality: Have the functional claims of the software been confirmed?
+
+[x] Performance: Have any performance claims of the software been confirmed?
+
+> N/A? I did not see any performance claims in README, I found
+> performance claims in
+> https://cran.r-project.org/web/packages/RcppCNPy/vignettes/RcppCNPy-intro.pdf
+
+### Documentation
+
+[x] A statement of need: Do the authors clearly state what problems
+the software is designed to solve and who the target audience is?
+
+> Target audience could be better stated in README and vignette
+
+[x] Installation instructions: Is there a clearly-stated list of
+dependencies? Ideally these should be handled with an automated
+package management solution.
+
+> I'm not sure if the authors would also like to include the list of
+> Python dependencies since Python example code is given in the README
+
+[x] Example usage: Do the authors include examples of how to use the
+software (ideally to solve real-world analysis problems).
+
+[x] Functionality documentation: Is the core functionality of the
+software documented to a satisfactory level (e.g. API method
+documentation)?
+
+> Functionality is better stated in vignette. README does not contain
+> exhaustive list of functionality
+
+[ ] Automated tests: Are there automated tests or manual steps
+described so that the function of the software can be verified?
+
+> Tests were performed, but I did not find a .md or README if that is
+> required by this journal
+
+[ ] Community guidelines: Are there clear guidelines for third parties
+wishing to 1) Contribute to the software 2) Report issues or problems
+with the software 3) Seek support
+
+> Community guidelines for contributions could be better reported
+
+### Software paper
+
+Paper PDF: 10.21105.joss.00055.pdf
+
+[x] Authors: Does the paper.md file include a list of authors with
+their affiliations?
+
+> Author names present in paper.md but did not compile to pdf
+
+[x] A statement of need: Do the authors clearly state what problems
+the software is designed to solve and who the target audience is?
+
+[x] References: Do all archival references that should have a DOI list
+one (e.g. papers, datasets, software)?
+
+> References do not have a DOI in pdf
+
+### Recommendations to editor
+
+While the software works well and quickly, some additional
+documentation is required by JOSS for publication.
+
+### Recommendations to author
+
+This reviewer would like to see reading and writing capability of .npz
+in the future. npyLoad of integer arrays produces incorrect values
+when not given the flag "integer." Multidimensional arrays should also
+be easily loadable. I was impressed that arrays of sizes (24,) ,
+(1,24), and (24,1) all loaded correctly in R with the existing
+functionality.
+
+### Recommendations to JOSS
+
+Reviewers should be left with the choice to be anonymous
