### 2016-09-21 #55 @mdnunez
[review.txt](https://github.com/openjournals/joss-reviews/files/485411/review.txt)

@eddelbuettel I've attached the markdown script
