### 2016-09-21 #55 @eddelbuettel
@mdnunez Very kind, thanks you -- I had already committed a [cobbled-together copy](https://github.com/eddelbuettel/rcppcnpy/blob/feature/joss/paper/joss/2016-09-20-review.md)  (NB: URL may disappear once I merge that branch).  Will probably update to yours.
