+++ paper/joss/2016-09-20-review.md
@@ -1,109 +1,86 @@
 # Review of RcppCNPy: Read-Write Support for NumPy Files in R
 
+
 ## Reviewer questions
 
 ### Conflict of interest
 
-[x] As the reviewer I confirm that there are no conflicts of interest
-for me to review this work (such as being a major contributor to the software).
+- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).
 
 ### General checks
 
-[x] Repository: Is the source code for this software available at the repository url?
+- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/eddelbuettel/rcppcnpy">repository url</a>?
 
 > Yes, it is also available from CRAN with install.packages('rcppcnpy')
 
-[x] License: Does the repository contain a plain-text LICENSE file
-with the contents of an OSI approved software license?
+- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
 
-> Yes, except it is GNU 2, not GNU 3 or later. Also, should GNU
-> statement be at the top of each script?
+> Yes, except it is GNU 2, not GNU 3 or later. Also, should GNU statement be at the top of each script?
 
-[x] Version: Does the release version given match the GitHub release (0.2.5)?
+- [x] **Version:** Does the release version given match the GitHub release (0.2.5)?
 
 ### Functionality
 
-[ ]  Installation: Does installation proceed as outlined in the documentation?
+- [ ] **Installation:** Does installation proceed as outlined in the documentation?
 
-> README does not list installation instructions, easy fix.  
-`library(rcppcnpy)` did not work for me  
-It seems `library('RcppCNPy')` command is case sensitive.
+> README does not list installation instructions, easy fix
+> library(rcppcnpy) did not work for me
+> It seems library('RcppCNPy') command is case sensitive.
 
-[x] Functionality: Have the functional claims of the software been confirmed?
+- [x] **Functionality:** Have the functional claims of the software been confirmed?
 
-[x] Performance: Have any performance claims of the software been confirmed?
+- [x] **Performance:** Have any performance claims of the software been confirmed?
 
-> N/A? I did not see any performance claims in README, I found
-> performance claims in
-> https://cran.r-project.org/web/packages/RcppCNPy/vignettes/RcppCNPy-intro.pdf
+> N/A? I did not see any performance claims in README, I found performance claims in https://cran.r-project.org/web/packages/RcppCNPy/vignettes/RcppCNPy-intro.pdf
 
 ### Documentation
 
-[x] A statement of need: Do the authors clearly state what problems
-the software is designed to solve and who the target audience is?
+- [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
 
 > Target audience could be better stated in README and vignette
 
-[x] Installation instructions: Is there a clearly-stated list of
-dependencies? Ideally these should be handled with an automated
-package management solution.
+- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
 
-> I'm not sure if the authors would also like to include the list of
-> Python dependencies since Python example code is given in the README
+> I'm not sure if the authors would also like to include the list of Python dependencies since Python example code is given in the README
 
-[x] Example usage: Do the authors include examples of how to use the
-software (ideally to solve real-world analysis problems).
+- [x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
 
-[x] Functionality documentation: Is the core functionality of the
-software documented to a satisfactory level (e.g. API method
-documentation)?
[…]
