### 2016-09-22 9fc4ad8
added feedback section and JOSS badge
