### 2016-09-22 b16a587
added doi for Rcpp reference
