+++ paper/paper.bib
@@ -5,7 +5,8 @@
   series =	 {Use R!},
   year =	 2013,
   address =	 {New York},
-  isbn =	 {978-1-4614-6867-7}
+  isbn =	 {978-1-4614-6867-7},
+  doi =          {10.1007/978-1-4614-6868-4}
 }
 
 @Manual{CRAN:Rcpp,
