### 2016-09-22 6d1586d
added examples section
