### 2016-09-22 #55 @mdnunez
It is in my view that @eddelbuettel has fulfilled all the requirements for publication in JOSS.

I feel the author has fulfilled the **Automated Test** requirement due to matching CRAN standards. However I await guidance from @cMadan to check this box as I'm not sure if JOSS requires written text markdown or README in addition to the included test script `tests/loadFiles.R`

I also noticed there is a minor mistake / redundancy in the example Python code in `README.md`:
```
>>> im = np.arange(3,4)
>>> im = np.arange(12).reshape(3,4)
```
