### 2016-09-22 7f43680
remove one redundant line
