### 2016-09-22 b845a3f
remove one redundant line
