### 2016-09-22 #55 @eddelbuettel
NB: there are actually two test scripts: `tests/saveAndLoad.R`, and `tests/loadFiles.R`.

And I removed the redundant Python line `np.arrange(3.4)`; that was a copy-and-paste left over from [the package webpage](http://dirk.eddelbuettel.com/code/rcpp.cnpy.html) where I removed it too.
