### 2016-09-22 86abf51
update ChangeLog and NEWS
