+++ inst/NEWS.Rd
@@ -2,6 +2,13 @@
 \title{News for Package \pkg{RcppCNPy}}
 \newcommand{\cpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
 
+\section{Changes in version 0.2.6 (2016-xx-yy)}{
+  \itemize{
+    \item Expanded documentation in README.md
+    \item Added examples to help page
+  }
+}
+    
 \section{Changes in version 0.2.5 (2016-08-26)}{
   \itemize{
     \item Synchronized code with the \code{cnpy} repository
