### 2016-09-22 #55 @eddelbuettel
Concerning @mdnunez concern about

> if JOSS requires written text markdown or README in addition to the included test script

I went back to the [Author guidelines](http://joss.theoj.org/about#author_guidelines) which are clearer:

> Tests

> Authors are strongly encouraged to include an automated test suite covering the core functionality of > their software.

> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to check the expected functionality of the  software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to check whether the software works

which we fully support in the package:
- unit tests automatic when the standard `R CMD check ....`  command is running,
- we have Travis CI connected (and all results [are available](https://travis-ci.org/eddelbuettel/rcppcnpy/builds)) to run checks on each commit or pull request.

I hope this provides additional clarification.   Thanks again for the prompt and detailed review.
