### 2016-09-22 #55 @mdnunez
@eddelbuettel gave convincing enough arguments to check the **Automated Test** box. (Btw: Are the Travis CI results linked to from the repository?)

No problem! This was overall a very pleasant review process.
