### 2016-09-22 #55 @cMadan
Thank you for the quick and thorough review, @mdnunez!

I have a few additional comments/suggestions before acceptance.

(1) paper.md didn't pull the author information properly (hopefully @arfon can look into that), however, the package vignette on CRAN (https://cran.r-project.org/web/packages/RcppCNPy/vignettes/RcppCNPy-intro.pdf) includes a second author, also listed in the README.md, that isn't included in the JOSS submission. They should probably also be included in paper.md.

(2) Since the package vignette in CRAN includes more information, it would be good to also reference/point-to that as well (in paper.md).

(3) README.md should have a small section for the installation instructions. I think it should include both the `install.packages("RcppCNPy")`, as well as `install.packages("eddelbuettel/rcppcnpy")`, should the CRAN package be a slightly older version than that found on GitHub.

(4) README.md should also include a sentence pointing to the `tests` directory. A short description of each test (in README.md or in a separate file) would also be beneficial.

(5) It would be good to cite NumPy in the paper, see https://www.scipy.org/citing.html.
