### 2016-09-22 #55 @eddelbuettel
Briefly

(1) I will add Wush.
(2) Yes, that occurred to me too, and I saw other JOSS submission do it. Will do.
(3) The delta is temporary and are the changes I made in the 24 hours in response to this very issue ticket / review.  I consider CRAN to be the primary and always push releases.  GitHub is my development repository.  This policy has been consistent and transparent for ten+ year across two dozen packages. I do not plan to change that.
(4) No. It is a CRAN package, and R users understand how to tickle tests.  The README in the repository is not a replacement for the R manuals.
(5) I look for such citation information, but found it wanting.  If anything, the [cnpy](https://github.com/rogersce/cnpy) repository should be cited (and I will add this).  RcppCNPy does not contain NumPy code or functionality so this is a bit of a stretch.   But I will add it.
