### 2016-09-22 #55 @cMadan
(3) I still think it would be good to add `install.packages("RcppCNPy")` to the README.md, to be overtly clear how the package can be installed from CRAN. This also may be useful in preventing confusion related to the case-sensitivity (also since the GitHub repo is named in all lowercase).

(5) Despite the package not using any NumPy code, it is built as a means to read-write from the NumPy data format--that's why I think citing NumPy makes sense. I hope the rationale is more clear now.
