### 2016-09-22 #55 @eddelbuettel
(3) Yes I agree (and I edited the above commented twice). I checked, and this package does not have the short block most my other packages have -- ie preference for `install.packages()`. I will add that.

(5) Yes. No point to debate this death.  Taken further, we could also cite R and Python. But NumPy is appropriate, and cnpy even more so.  Any tips on citing 'unpublished' github repos?
