### 2016-09-22 0ec8055
updates to paper and bib
awaiting possible orcid id from Wush
