+++ paper/paper.bib
@@ -15,6 +15,15 @@
                   Allaire and Kevin Ushey and Qiang Kou and
                   John Chambers and Douglas Bates},
   year =	 2016,
-  note =	 {R package version 0.12.6},
+  note =	 {R package version 0.12.7},
   url =		 {http://CRAN.R-Project.org/package=Rcpp}
+}
+
+@Misc{NumPy,
+  author =       {Eric Jones and Travis Oliphant and Pearu Peterson
+                  and others},
+  title =        {{SciPy}: Open source scientific tools for {Python}},
+  year =         {2001--2016},
+  url =          "http://www.scipy.org/",
+  note =         {[Online; accessed 2016-09-22]}
 }
