### 2016-09-22 #55 @eddelbuettel
@arfon: Regarding `paper.md` and `paper.pdf`, I raise a white flag. I must be using the wrong pandoc invocation. I even compared carefully against the files in two already-published JOSS pieces.

I have used

```bash
$ pandoc paper.md --output paper.pdf --filter pandoc-citeproc
```

without additional bells and whistles.  Do I need some?    Pandoc is 1.16.0.2 on Ubuntu 16.04.
