### 2016-09-22 #55 @cMadan
For citing unpublished GitHub repos, take a look at this:
http://academia.stackexchange.com/a/14015

As for compiling paper.md, normally @arfon does this (not the author). After you commit these last changes, I'll be happy to accept your submission for publication in JOSS.
