### 2016-09-23 7a7af09
expanded .bib once more
