+++ paper/paper.bib
@@ -16,8 +16,17 @@
                   John Chambers and Douglas Bates},
   year =	 2016,
   note =	 {R package version 0.12.7},
-  url =		 {http://CRAN.R-Project.org/package=Rcpp}
+  url =		 {https://CRAN.R-Project.org/package=Rcpp}
 }
+
+@Manual{CRAN:RcppCNPy,
+  title =	 {RcppCNPy: DRead-Write Support for 'NumPy' Files via 'Rcpp'},
+  author =	 {Dirk Eddelbuettel and Wush Wu},
+  year =	 2016,
+  note =	 {R package version 0.2.5},
+  url =		 {https://CRAN.R-Project.org/package=RcppCNPy}
+}
+
 
 @Misc{NumPy,
   author =       {Eric Jones and Travis Oliphant and Pearu Peterson
@@ -26,4 +35,14 @@
   year =         {2001--2016},
   url =          "http://www.scipy.org/",
   note =         {[Online; accessed 2016-09-22]}
+}
+
+@misc{CNPy,
+  author =       {Carl Rogers},
+  title =        {cnpy},
+  year =         2015,
+  publisher =    {GitHub},
+  journal =      {GitHub repository},
+  howpublished = {\url{https://github.com/rogersce/cnpy}},
+  commit =       {4f57d6a0e4c030202a07a60bc1bb1ed1544bf679}
 }
