### 2016-09-23 #55 @eddelbuettel
Yes, agreed, `@Misc` or `@Manual` are good BibTeX types for that. I just updated the .bib and .pdf [once more](https://github.com/eddelbuettel/rcppcnpy/commit/7a7af0908ab6f72872e84d8e42b9254eb91ec9fd). I also added a reference for the actual package RcppCNPy.

More to come, will report back once done.
