### 2016-09-23 #55 @cMadan
Sounds good!

Just since I took a look, typo "DRead" in the bib. For NumPy, it would be better to cite the suggested 2011 paper (https://www.scipy.org/citing.html), rather than the SciPy website.
