### 2016-09-23 d590b5c
correct NumPy reference
