+++ paper/paper.bib
@@ -20,21 +20,25 @@
 }
 
 @Manual{CRAN:RcppCNPy,
-  title =	 {RcppCNPy: DRead-Write Support for 'NumPy' Files via 'Rcpp'},
+  title =	 {RcppCNPy: Read-Write Support for 'NumPy' Files via 'Rcpp'},
   author =	 {Dirk Eddelbuettel and Wush Wu},
   year =	 2016,
   note =	 {R package version 0.2.5},
   url =		 {https://CRAN.R-Project.org/package=RcppCNPy}
 }
 
-
-@Misc{NumPy,
-  author =       {Eric Jones and Travis Oliphant and Pearu Peterson
-                  and others},
-  title =        {{SciPy}: Open source scientific tools for {Python}},
-  year =         {2001--2016},
-  url =          "http://www.scipy.org/",
-  note =         {[Online; accessed 2016-09-22]}
+@Article{NumPy,
+  author =       {Stéfan {van der Walt} and S. Chris Colbert and
+                  Gaël Varoquaux},
+  title =        {The NumPy Array: A Structure for Efficient Numerical
+                  Computation},
+  journal =      {Computing in Science & Engineering},
+  year =         2011,
+  volume =       13,
+  number =       2,
+  pages =        {22-30},
+  url =          {http://scitation.aip.org/content/aip/journal/cise/13/2/10.1109/MCSE.2011.37},
+  doi =          {10.1109/MCSE.2011.37}
 }
 
 @misc{CNPy,
