### 2016-09-23 #55 @eddelbuettel
Typo fixed.  NumPy reference fixed; thanks for catching that. It's been a long day...
