### 2016-09-23 8e57daf
add installation note to README, expand feeback
