### 2016-09-23 #55 @eddelbuettel
I just expanded `README.md` as discussed, and with that should have addressed the points you made yesterday.   All changes are still in the [features/joss branch](https://github.com/eddelbuettel/rcppcnpy/tree/feature/joss) which I intend to merge once we have finalized the review.
