### 2016-09-23 #55 @cMadan
Perfect, I am happy to accept the submission in it's current form.

@arfon - for archiving the code, does CRAN count as a suitable repository or should it also be in zenodo/figshare ?
