### 2016-09-23 #55 @arfon
> @arfon - for archiving the code, does CRAN count as a suitable repository or should it also be in zenodo/figshare ?

No, it also needs to be in Zenodo or figshare.
