### 2016-09-24 #55 @eddelbuettel
I have resisted Zenodo so far because the 'one id per release' seems "wrong" -- for some active packages of mine (such as Rcpp or RcppArmadillo) we have six or more releases a year, every year.

Do you want me to set the Zenodo id up for release 0.2.5 that the submission was based on?

Also, I now have the orcid id of my coauthor so there will be one more commit.
