### 2016-09-24 #55 @eddelbuettel
Now done in [this commit](https://github.com/eddelbuettel/rcppcnpy/commit/885fcda79628be85cb43f1fe059f675b9b7177a2).
