### 2016-09-24 #55 @cMadan
Thanks for adding the orcid!

I would say the Zenodo archive doi should correspond to the 'current' version. I.e., version 0.2.5, or newer if you are incrementing the version number based on the changes since the initial submission.
