### 2016-09-24 #55 @eddelbuettel
Well I was planning to finalize this submission, review and publication (submitted as 0.2.5).  I would then update the package, include the paper and metadata etc pp and call that 0.2.6.
