### 2016-09-24 #55 @eddelbuettel
I created a Zenodo doi:  [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.155066.svg)](https://doi.org/10.5281/zenodo.155066)
