### 2016-09-24 e1761f2
added zenodo doi to README.md
