### 2016-09-24 #55 @arfon
@eddelbuettel thanks. Are you planning on making a final Zenodo release with all of the changes (such as including the ORCID of your coauthor) or does [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.155066.svg)](https://doi.org/10.5281/zenodo.155066) already include these changes?
