### 2016-09-24 #55 @eddelbuettel
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.155066.svg)](https://doi.org/10.5281/zenodo.155066) archives the current CRAN release 0.2.5 on which the submission was based.

I plan to release 0.2.6 once we are done here and could provide another Zenodo DOI for that.
