### 2016-09-25 #55 @arfon
> I plan to release 0.2.6 once we are done here with all the changes in `paper/` and other files which accumulated in the review privess, and could provide another Zenodo DOI for that.

@eddelbuettel - based on @cMadan's [comment](https://github.com/openjournals/joss-reviews/issues/55#issuecomment-249260290) I think we're good to accept in the current state. If you want to make the 0.2.6 release and the corresponding Zenodo archive I can move forward with accepting the submission.
