### 2016-09-25 #55 @cMadan
Right, I do think 0.2.6 should be the accepted version, since that's the version that includes the changes made as part of the review.
