### 2016-09-25 #55 @eddelbuettel
Sounds good, and as I recall that is also how JSS does it.  So what if any updated files (or information) do I need from you to finalize 0.2.6?
