### 2016-09-25 #55 @arfon
> In other words: I am planning on rolling out 0.2.6 once we are done here. So we appear to be in a deadlock: you are waiting for me, I am waiting for you. All changes I made are in the feature/joss branch.

I believe you can move forward and make the 0.2.6 release now based on the `feature/joss` branch.
