### 2016-09-25 #55 @eddelbuettel
No -- I cannot reference a JOSS publication.
