### 2016-09-25 #55 @arfon
> No -- I cannot reference a JOSS publication.

Sorry, where do you need to create this reference?
