### 2016-09-25 #55 @eddelbuettel
It would be part of the post-publication release and tarball, including inter alia a file `inst/CITATION` providing data for R's `citation()` function.
