### 2016-09-25 #55 @arfon
Ah OK. Well, the DOI for the paper will be http://dx.doi.org/10.21105/joss.00055 (although I've not created the DOI yet). Can you just include that?
