### 2016-09-25 #55 @eddelbuettel
Yes I could.  Are you saying that you want me to _blink first_ so that you can finalize, ie you need 0.26 present?

(Also see my last round of edits to prev comment)
