### 2016-09-25 #55 @arfon
> Yes I could. Are you saying that you want me to blink first so that you can finalize, ie you need 0.26 present?

😀 that would be great. The thing I really need is the DOI pointing to the 0.2.6 release.

> The whole thing is weird

I agree this is a tad strange and we're still working through some of these issues in our editorial process. Thanks for your patience :-)
