### 2016-09-25 #55 @eddelbuettel
Let me give it a look over in the morning.
