### 2016-09-25 817ce56
added changelog and news
