+++ inst/NEWS.Rd
@@ -2,10 +2,11 @@
 \title{News for Package \pkg{RcppCNPy}}
 \newcommand{\cpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
 
-\section{Changes in version 0.2.6 (2016-xx-yy)}{
+\section{Changes in version 0.2.6 (2016-09-xx)}{
   \itemize{
     \item Expanded documentation in README.md
     \item Added examples to help page
+    \item Added CITATION file for JOSS paper
   }
 }
     
