### 2016-09-25 #55 @eddelbuettel
Here is one thing I certainly need from you:  volume issue, as well as a look over.  File `inst/CITATION` currently has

```r
citHeader("To cite RcppCNPy in publications use:")

citEntry(entry = "Article",
  title        = "{RcppCNPy}: Read-Write Support for NumPy Files in R",
  author       = personList(as.person("Dirk Eddelbuettel"), as.person("Wush Wu")),
  journal      = "Journal of Open Source Software",
  publisher    = "The Open Journal",
  year         = "2016",
  volume       = "1",
  number       = "55"
  month        = "September",
  url          = "http://dx.doi.org/10.21105/joss.00055",

  textVersion  =
  paste("Dirk Eddelbuettel, Wush Wu (2016).",
        "RcppCNPy: Read-Write Support for NumPy Files in R.",
        "Journal of Open Source Software, 1(55), September 2016.",
        "URL http://dx.doi.org/10.21105/joss.00055")
)
```

Month, volume, issue are probably wrong. I hand-formated the text variant. do you want changes?

I had a `doi` entry in there, but then R chokes over it when running `citation("RcppCNPy")`.  As shown above, it works (but may need tweaks):

```r

R> citation("RcppCNPy")

To cite RcppCNPy in publications use:

  Dirk Eddelbuettel, Wush Wu (2016). RcppCNPy: Read-Write Support for NumPy Files in R.
  Journal of Open Source Software, 1(55), September 2016.
  URL http://dx.doi.org/10.21105/joss.00055

A BibTeX entry for LaTeX users is

  @Article{,
    title = {{RcppCNPy}: Read-Write Support for NumPy Files in R},
    author = {Dirk Eddelbuettel and Wush Wu},
    journal = {Journal of Open Source Software},
    publisher = {The Open Journal},
    year = {2016},
    volume = {1},
    number = {55},
    month = {September},
    url = {http://dx.doi.org/10.21105/joss.00055},
  }

R>
```
