### 2016-09-25 #55 @eddelbuettel
Last but not least you would have to turn the publication on so that I can refer to it:

```
DOI Not Found

10.21105/joss.00055

This DOI cannot be found in the DOI System. Possible reasons are:

The DOI is incorrect in your source. Search for the item by name, title, or other metadata using a search engine.
The DOI was copied incorrectly. Check to see that the string includes all the characters before and after the slash and no sentence punctuation marks.
The DOI has not been activated yet. Please try again later, and report the problem if the error continues.
```
