### 2016-09-25 5da0a4e
use canonical URLs
