### 2016-09-25 #55 @eddelbuettel
Which would block CRAN admission:

```bash
edd@max:~/git$ RD CMD check --as-cran RcppCNPy_0.2.6.tar.gz      ## RD launches R-devel
* using log directory ‘/home/edd/git/RcppCNPy.Rcheck’
* using R Under development (unstable) (2016-09-20 r71314)
* using platform: x86_64-pc-linux-gnu (64-bit)
* using session charset: UTF-8
* using option ‘--as-cran’
* checking for file ‘RcppCNPy/DESCRIPTION’ ... OK
* checking extension type ... Package
* this is package ‘RcppCNPy’ version ‘0.2.6’
* checking CRAN incoming feasibility ... NOTE
Maintainer: ‘Dirk Eddelbuettel <edd@debian.org>’

Found the following (possibly) invalid URLs:
  URL: http://dx.doi.org/10.21105/joss.00055
    From: inst/CITATION
    Status: 404
    Message: Not Found
* checking package namespace information ... OK
* checking package dependencies ... OK
[....]
```
