### 2016-09-25 #55 @arfon
OK, the DOI is registered now: http://dx.doi.org/10.21105/joss.00055 but it won't resolve to the JOSS site properly yet.

Quoting from the submitted Crossref metadata:

```
      <journal_issue>
        <publication_date media_type="online">
          <month>09</month>
          <year>2016</year>
        </publication_date>
        <journal_volume>
          <volume>1</volume>
        </journal_volume>
        <issue>5</issue>
      </journal_issue>
```

Also, please merge [this PR](https://github.com/eddelbuettel/rcppcnpy/pull/10) into your changes before making the 0.2.6 release. The current paper metadata is malformed.
