### 2016-09-25 #55 @eddelbuettel
Well, close, but not cigar:

```bash
* this is package ‘RcppCNPy’ version ‘0.2.6’
* checking CRAN incoming feasibility ... NOTE
Maintainer: ‘Dirk Eddelbuettel <edd@debian.org>’

Found the following (possibly) invalid URLs:
  URL: http://dx.doi.org/10.21105/joss.00055
    From: inst/CITATION
    Status: 500
    Message: Internal Server Error
* checking package namespace informati
```
