### 2016-09-25 #55 @eddelbuettel
Volume and issue updated [in this commit to branch](https://github.com/eddelbuettel/rcppcnpy/commit/69b0d396237be15d699af2305bdc364bd72a3bbf) and now yields:

```r
R> citation("RcppCNPy")

To cite RcppCNPy in publications use:

  Dirk Eddelbuettel, Wush Wu (2016). RcppCNPy: Read-Write Support for NumPy Files in R.
  Journal of Open Source Software, 1(5), September 2016.
  URL http://dx.doi.org/10.21105/joss.00055

A BibTeX entry for LaTeX users is

  @Article{,
    title = {{RcppCNPy}: Read-Write Support for NumPy Files in R},
    author = {Dirk Eddelbuettel and Wush Wu},
    journal = {Journal of Open Source Software},
    publisher = {The Open Journal},
    year = {2016},
    volume = {1},
    issue = {5},
    month = {09},
    url = {http://dx.doi.org/10.21105/joss.00055},
  }

R>
