### 2016-09-25 #55 @arfon
> Well, close, but no cigar:

Can you produce the archive (Zenodo/figshare) of the 0.2.6 release _before_ uploading to CRAN? That way we won't run into this:

```
Found the following (possibly) invalid URLs:
  URL: http://dx.doi.org/10.21105/joss.00055
    From: inst/CITATION
    Status: 500
    Message: Internal Server Error
* checking package namespace information ... OK
* checking package dependencies ... OK
```
