### 2016-09-25 #55 @arfon
> Also, please merge this PR into your changes before making the 0.2.6 release. The current paper metadata is malformed.

Also, https://github.com/eddelbuettel/rcppcnpy/pull/10 isn't required anymore. It looks like you fixed this in your `feature/joss` branch.
