### 2016-09-25 #55 @eddelbuettel
Not really.   I could prepare a _candidate_ tarball, but CRAN is at least as picky as you, and there may be revisions leading to changes in the tarball invalidating the Zenodo doi.

The easiest, really, would be to point to the _exiting_ Zenodo doi for 0.2.5.  Version 0.2.6 will go out, as will (in all likelihood) versions 0.2.7, 0.2.8 and so on.  So there is always a delta.  Why not seal things now?
