### 2016-09-25 #55 @arfon
Ok I give up. Let's do that.

If ever there was a case for versioned DOIs this is it. 😂
