### 2016-09-25 #55 @eddelbuettel
> Ok I give up. Let's do that.

Woot!

> If ever there was a case for versioned DOIs this is it. :joy:

I have been thinking about this too.   There is a conflict between software changing and updating as it is never final -- so a DOI for the master branch or its HEAD ? -- as it would not make real sense to create, ie, six DOIs a year for the six Rcpp releases I make.   Yet the one literature reference (the original JSS paper) is really stale.   I'd love a citable reference for 'current Rcpp' (or any other changing software package).  Not clear.   And of course in conflict with the idea of "perfectly reproducible research" and snapshots (but that is a whole other discussion as there are some issues too...

Off for a run now, at last.  Let me mull this over one more time.   But thanks for moving on along...
