### 2016-09-25 #55 @arfon
> but CRAN is at least as picky as you

Also, for what it's worth, I don't think it's unreasonable (or that picky) to ask authors to prepare an archive of the software that includes the changes as a result of the JOSS review but this is clearly a problem for folks (like yourself) who are submitting to CRAN.

/ cc @karthik for any ideas how to make this easier in the future.
