### 2016-09-25 #55 @arfon
> Off for a run now, at last. Let me mull this over one more time. But thanks for moving on along...

OK thanks. I'll move forward with accepting.
