### 2016-09-25 #55 @arfon
@mdnunez - many thanks for the review!

@eddelbuettel - your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00055 🎉 🚀 💥
